Refactor executors and create OneInExecutor
